190410 | login/logout

+++ django/insta/posts/views.py
@@ -1,12 +1,15 @@
 from django.shortcuts import render, redirect, get_object_or_404
+from django.contrib.auth.decorators import login_required
 from .forms import PostForm
 from .models import Post
 
 
 def list(request):
-    posts = Post.objects.all()
+    posts = Post.objects.order_by('-id').all()
     return render(request, 'posts/list.html', {'posts': posts})
 
+
+@login_required
 # request: 외부의 요청이 담긴 정보
 def create(request):
     if request.method == 'POST':
@@ -14,15 +17,21 @@ def create(request):
         #data = request.POST, file = request.FILES 생략 되있는 형태지만 자리 맞추면 인자가 고대로 들어감
         post_form = PostForm(request.POST, request.FILES)
         if post_form.is_valid():
-            post_form.save()
+            post = post_form.save(commit=False)
+            post.user = request.user
+            post.save() # 실제 데이터베이스에 저장
             return redirect('posts:list')
     else:
         post_form = PostForm()
     
-    return render(request, 'posts/create.html', {'post_form': post_form})
+    return render(request, 'posts/form.html', {'post_form': post_form})
     
 def update(request, post_id):
     post = get_object_or_404(Post, pk=post_id)
+    
+    if post.user != request.user:
+        return redirect('posts:list')
+    
     if request.method == 'POST':
         post_form = PostForm(request.POST, request.FILES, instance=post)
         if post_form.is_valid():
@@ -30,7 +39,7 @@ def update(request, post_id):
             return redirect('posts:list')
     else:
         post_form = PostForm(instance=post)
-    return render(request, 'posts/create.html',{'post_form': post_form})
+    return render(request, 'posts/form.html',{'post_form': post_form})
 
 #삭제
 # variable routing => post_id
@@ -39,6 +48,10 @@ def delete(request, post_id):
     # post = Post.objects.get(pk=post_id)
     # 지운걸 또 지우려 할 떄 존재하지 않는다고 알려줌
     post = get_object_or_404(Post, pk=post_id)
+    
+    if post.user != request.user:
+        return redirect('posts:list')
+        
     post.delete()
     
     return redirect('posts:list')
